# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Catalogo de vinos",
        "description": "Descripción de mi API",
        "termsOfService": "https://example.com/terms/",
        "contact": {
            "email": "jesquuliche@hotmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/tipo": {
            "get": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Get list of tipos",
                "description": "Returns list of tipos",
                "operationId": "getTiposList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nombre": {
                                                "type": "string",
                                                "example": "Tipo 1"
                                            },
                                            "descripcion": {
                                                "type": "string",
                                                "example": "Descripción del tipo 1"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Store a new tipo",
                "description": "Stores a new tipo and returns it",
                "operationId": "storeTipo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Nuevo Tipo"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción del nuevo tipo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Nuevo Tipo"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción del nuevo tipo"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tipo/{id}": {
            "get": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Get tipo by ID",
                "description": "Returns a single tipo",
                "operationId": "getTipoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of tipo to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Tipo 1"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción del tipo 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Update an existing tipo",
                "description": "Updates an existing tipo and returns it",
                "operationId": "updateTipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of tipo to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Tipo Actualizado"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Descripción del tipo actualizado"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Tipo Actualizado"
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "Descripción del tipo actualizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tipo not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tipos"
                ],
                "summary": "Delete a tipo",
                "description": "Deletes a single tipo",
                "operationId": "deleteTipo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of tipo to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Tipo not found"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}